requirements.txt
openai
gtts
moviepy
python-dotenv
